refactor: Car 객체 domain 분류
